Links extraction: skip dfn panels (#1653)
Via https://github.com/w3c/reffy/issues/1584#issuecomment-2304730538.

Aside dfn panels only contain links that already appear somewhere else in the
spec.

These links were extracted and categorized as `autolinks` when they appeared in
the final index of a Bikeshed spec, which essentially did not do anything
because the links had already been extracted as `autolinks` in any case.

The problem is that Respec also uses similar panels (but no `<aside>`), and
Bikeshed sometimes outputs the panels at the end of the document, and not
within the index itself as expected in the code. In these situations, the
links were extracted a second time, as `rawlinks`. Not extracting the panels
avoids duplication entirely.