Normalize newlines when slicing IDL extracts text
IDL extracts are generated using `\n` as newline delimiter. However,
platform-specific newlines may be introduced during curation when patches are
applied, meaning that, on Windows, curated extracts may end up with a mix of
`\n` and `\r\n` characters. This confused the code that read the extracts and
dropped the header.

The code now normalizes all newlines to `\n` before it slices the content.