Support CDDL dfns, extract standard prelude from CDDL RFC (#1750)
New types of definitions will soon be introduced in Bikeshed to deal with
CDDL blocks: cddl-module, cddl-type, cddl-parameter, cddl-key, cddl-value. The
extraction now considers that these definition types are valid.

To be able to reference CDDL types defined in the standard prelude, a
pre-processing step is also added for the CDDL RFC (RFC8610). To avoid creating
multiple definitions with the same ID and href, which could perhaps confuse
consuming tools (?), all types defined in the standard prelude get attached as
different linking text to the same cddl-type dfn.