Drop cache info when a redirection took place (#1777)
When a redirection takes place, Reffy follows the redirection logic (typically
done through scripting) but the cache info it gets from Puppeteer remains for
the initial URL. Reffy incorrectly assumed that info also applied to the final
page.

There's no easy way to retrieve the cache info of the final URL. Since that
should only affect a spec that moves, and only until we detect and update the
URL of the spec in browser-specs, this update simply drops the cache info to
force Reffy to crawl the spec.

Fix #1774.