Dfns extraction: expand logic that skips imported definitions (#1789)
The crawler expected "imported definitions" to use `<dfn>` as the wrapping
element with an inner `<a>` element. The Source map format spec used the
opposite pattern. This update expands the logic accordingly.

The Source map format spec actually swapped back the elements, but there
are a couple of other specs that use the same pattern (Audiobooks and the
JSON-LD API), and more specs may elect to do that in the future as well.

Close #1788.