Add `--markdown` and `--summary` crawl options (#1811)
The `--summary` crawl option tells Reffy to attach a crawl summary in Markdown
to each spec result. The summary is added to a `crawlSummary` property.

The `--markdown` crawl option tells Reffy to output a Markdown report instead
of the regular JSON output.

The Markdown summary is fairly basic for now. It starts with a short summary,
then details interesting things that got extracted (CSS terms, exported terms,
Web IDL), along with links to the online xref database search where
appropriate. The details are reported in expandable sections.

The summary is very likely to evolve:
- to get additional details on other extracts (such as CDDL or headings)
- to improve the ordering and formatting (for example, details on exported
terms could probably omit IDL terms that are "for" something since that's less
likely to be a nuisance, and could probably omit IDL terms altogether as main
terms get reported in the IDL details)

The options will allow us to report on crawl results in browser-specs issues
when a new spec gets proposed.