Markdown report: improve dfns, fix links, sort summary
The markdown report now separates between dfns that are exported by default
(such as CSS and Web IDL terms) and those that editors explicitly exported.
Details only report on the explicitly exported terms both because they should
be the most interesting ones to review, and because we don't want to create
too long a report for usual specs, as GitHub imposes constraints to the length
of an issue comment.

This update sorts the summary to ease reading.

The logic that handled links was also incorrect.