Authorize alternateIds property in dfns schema as well (#1843)
Close #1842.

Note: the update does not attempt to move the definition of a heading to be
common between dfns extracts and headings extracts because dfns extraction
can produce headings that slightly deviate from the structure used in headings
extracts:
1. headings in dfns extracts may only consist of an `href` and `title`, to
indicate that the dfn was found outside of any heading (not that this happens
much outside of the context of tests).
2. headings in dfns extracts do not have information about `level` for now.