Move explicit test timeout from suite to test level (#1846)
There seems to be an important nuance between Mocha and Node.js with regards
to timeouts: a call to `this.timeout()` in Mocha at the suite level says
"Apply the timeout to individual tests", whereas specifying the timeout at
the suite level in Node.js seems to say "Apply the timeout to individual
tests **and** to the entire suite".

I'm getting timeouts when I run tests locally as a result (although that
does not seem to affect tests execution in GitHub actions for now). This
update moves the timeout to the test level where it matters. There are
other places where the timeout is set at the suite level, but these typically
contain only one test, so no need to move.