Set selector syntax to selector name when possible
The syntax of selectors is often not formally defined in specs. For selectors
that are function-like such as `:nth-child()`, we'll have to see how specs can
be updated. For selectors are are keyword-like such as `:visited`, updating the
specs seems more confusing than anything else though: the syntax of the
selector goes without saying, it is simply the selector's name.

This update makes Reffy set the selector's syntax to be the selector's name
when the selector is keyword-like.