Add cssmerge post-processor to consolidate CSS extracts
This adds a cssmerge post-processor at the crawl level that consolidates
definitions from CSS extracts into a single file. The CSS entries in the
resulting file are de-duplicated. The module starts with a number of comments
that detail the approach.

The intent is to create a view similar to that offered in MDN data to address
https://github.com/w3c/webref/issues/1519

Comments at the top of the module also detail the main differences between the
two views. Main one is structural: we tend to use arrays in Webref so this code
also produces arrays, while MDN data uses indexed objects. It's easy to build
an indexed view from an array though.

Consolidation focuses on syntaxes and the code does not attempt to compute the
`units` list that exists in MDN data. It is easy to build that list by looking
at values of a handful of CSS properties. We could add it later on to the
consolidated file if that proves needed.

A couple of remaining open questions:
1. This creates a `css.json` file, similar to the `events.json` file that we
create when we post-process events. That's nice but it would be good to ship
that file in `@webref/css` in the end. That package already has a `CSS.json`
file at the root. How/Where to package the `css.json` file?
2. To de-duplicate, the most recent definition is used. That's the only
approach I can think of that can be automated. Consumers of MDN data may need
a more nuanced approach where the syntax better matches what core browsers
currently support. We can perhaps refine the process later on if needed with
some sort of patching mechanism (but my hunch is that this will remain a manual
process).