[CSS post-processing] Merge scopes when possible (breaking)
A CSS feature may be defined within a single dfn for multiple scopes. These
features appeared in separate entries in the consolidated file. To ease
indexing and lookup by feature name, this update further merges scopes.

This does not fully turn feature names into identifiers, but only 8 features
remain that use the same name but have different definitions for different
scopes. Some (but not all) of them, we should be able to get rid of through
spec fixes and possible improvements to the curation logic in Webref.

See list and details at:
https://github.com/w3c/webref/issues/1519#issuecomment-2962111268

This is a breaking change since it turns `for` keys into arrays in the
consolidated file.

Side note: the update also fixes an unrelated comment in the code.