Represent dev anchors in a `links` property
Instead of creating a separate definition, this records anchors for developers
in a `links` property attached to the base definition that the anchor
describes, as suggested in:
https://github.com/w3c/reffy/pull/1875#issuecomment-3045441655

The `links` property is an array of objects, each object featuring `id`,
`href`, `type`, `name` and `heading` properties. The `type` property is always
set to `"dev"`. The `name` property contains the text content of the enclosing
`<dt>`. The `heading` property contains the heading of the section where the
anchor is defined (it may be different from the heading of the section where
the underlying definition appears).

Notes:
- This approach avoids having to create a separate namespace for anchor
definitions.
- There may be more than one dev link per definition. That's normal. It
typically happens when the underlying definition is for a mixin included in
multiple interfaces, as for `TextDecoderCommon` attributes in the Encoding
spec.
- Some anchors for developers target definitions in external specs. Previous
code ignored them but could have added them as separate entry if need had
arisen. There is no easy way to add these anchors with the new approach if
they turned out to be useful.