Add anchors for web developers to dfns extracts (#1875)
For context, see discussion starting at:
https://github.com/mdn/browser-compat-data/pull/23958#issuecomment-3035921128

Some specs such as DOM, Encoding, HTML contain sections targeted at web
developers. These sections re-define terms normatively defined elsewhere in a
more developer-friendly way. Terms re-defined in these sections are good
targets for documentation but did not appear in definitions extracts.

This update makes Reffy parse "for web developers" sections and extract the
links that complete definitions they contain. This is a prerequisite to publishing
a package with definitions that could be used to validate URLs in BCD and
web-features, as envisioned in:
https://github.com/w3c/webref/issues/1198#issuecomment-3032925937

The links are recorded in a `links` property attached to the base definition that
the link completes. The `links` property is an array of objects, each object
featuring `id`, `href`, `type`, `name` and `heading` properties. The `type` property
is always set to `"dev"`. The `name` property contains the text content of the
enclosing `<dt>`. The `heading` property contains the heading of the section
where the anchor is defined (it may be different from the heading of the section
where the underlying definition appears).

There may be more than one dev link per definition. That's normal. It typically
happens when the underlying definition is for a mixin included in multiple
interfaces, as for `TextDecoderCommon` attributes in the Encoding spec.

Some links for developers target definitions in external specs. They are ignored
for now.

Worth noting:
- Ideally, spec authoring tools would provide better support for this pattern,
giving these links more stable IDs than `ref-for-[foo][number]` and possibly
creating proper dfns themselves. If they do that, processing may need to be
adjusted. Updating tools and specs will take time though.
- The key marker for sections targeted at web developers is the use of a
`domintro` class. Now, a few specs do use `domintro` in normative definition
lists (shape-detection-api, image-capture, mediastream-recording). That's
probably unintentional. I'll look into fixing the specs. The code skips
`domintro` sections that look suspicious.
- This would add **2815 links** to the dfns extracts (for ~50000 definitions)