Extract production rules dfns from ECMAScript spec (#1881)
Via https://github.com/mdn/browser-compat-data/pull/27246

This completes preprocessing of the ECMAScript spec to also extract production
rules that are needed for documentation purpose. These production rules are
defined in `<emu-production>` tags wrapped in `<emu-grammar type="definition">`
tags.

Annex B re-defines some of the production rules for Web browser hosts. To
avoid duplication, the code adds a `"for": ["Web browsers"]` to disambiguate.

All definitions are created with a "grammar" definition type, and flagged as
"private" (at least until it becomes clear that the terms are needed for xref
purpose). Use of "grammar" is not fantastic but is already done by the CSP
spec and it avoids introducing a new definition type that isn't explicitly
supported by spec authoring tools.