Force add CSS functions/types that do not have a proper dfn
When a function or type was found in a production rule that does not have a
`<dfn>`, it was added to the list of "missing definitions". In practice, the
only cases where that happens are:
- When a spec extends a construct defined elsewhere, see #1647. In such cases,
we'd want the extended type definition to appear in the extract.
- For `<boolean-expr-group>`, see #1878. In such cases, we'd rather not add
the type since the underlying spec needs fixing, but that can be caught
automatically, see below.

This code forces the creation of a function/type. This is breaking change
because, up until now, we guaranteed that all entries would have an `href` and
that will no longer be the case.

Ideally, we'd restrict this logic to extensions, but production rules are
processed as raw text, meaning we don't have the link information that would
tell us that we're looking at an extension. That is not such a big deal though:
it is easy to tell when an entry without an `href` extends another, and it
seems fine to use patching to curate the data (only one occurrence today).

CSS consolidation is updated to deal with extended functions and types. As
opposed to extensions of properties where new values are added to the syntax,
function/type extensions re-define the entire syntax.