Add custom dfns/headings extraction logic for WebGL1
The HTML of the WebGL1 spec was mostly written by hand. It contains a few
actual definitions that follow the definitions data model, but it also defines
~15+8 anchors for IDL attributes and methods that do not. These anchors are
typically used in BCD.

A new pre-processing step was added to turn these anchors into proper `<dfn>`
tags in WebGL1. As with SVG 2, that pre-processing step needs to parse the IDL
to scope definitions properly (the prose is clunky, see inline comments in the
code).

Similarly, a couple of headings have an additional anchor that is more
human-friendly than `5.2` and that BCD typically targets. These IDs are now
captured in the `alternateIds` array.