[CSS consolidation] Only keep most recent descriptor definition (#1900)
Consolidation already tried to discard descriptors when a newer definition
exists, but the code always kept the descriptors from the base definition in
practice. The code now also filters descriptors from the base definition as
needed.

This update also drops the `type` key for functions and types in the
consolidated file per suggestion in:
https://github.com/w3c/webref/issues/1519#issuecomment-3043456761
(the key was useless given that functions and types already appear in distinct
categories)