[CSS consolidation] Add `extended` to store list of extension URLs (#1904)
This adds a new `extended` key to entries in the consolidated CSS that lists
URLs of specs that extend the base definition. For properties, the URLs target
the actual extended definition. For functions and types, the URLs merely target
the spec without any specific fragment, because there is no actual definition
and we don't know where the extension appears at the consolidation phase.

Note this also prepares the consolidation to deal with extended functions and
types. As opposed to extensions of properties where new values are added to the
syntax, function/type extensions re-define the entire syntax.

Reffy does not support extraction of function/type extensions in itself yet, so
that part of the update is not going to be used for now.