Repair and extent refs extraction for ECMA specs (#1924)
Normative references listed in the ECMAScript spec were no longer being
extracted because the spec now lists them using `<p>` tags.

This update also adds support for the very similar format used in the Source
map format specification (closes #1792). That spec notes in an HTML comment
that the informative references should be normative references. We extract
these references as informative, which seems good as we shouldn't have to
rely on an HTML comment hidden in the source...