CSS merge: flatten nested at-rules (breaking change) (#1962)
The `@font-feature-values` at-rule defined in `css-fonts-4` defines nested
at-rules. These were captured as descriptors and did not appear at the root
level as scoped at-rules. This wasn't consistent with the way scoped functions
and types are handled in the `@webref/css` package.

This update flattens the at-rules, scoping them to the at-rule that they are
defined for.

Note: for extra-consistency with scoped functions and types, the `for` property
is defined as an array.

There are 7 scoped at-rules, all scoped to `@font-feature-values`:
- `@annotation`
- `@character-variant`
- `@historical-forms`
- `@ornaments`
- `@styleset`
- `@stylistic`
- `@swash`

This is a breaking change in that there wasn't any scoped at-rule until now and
the `for` property was not allowed for at-rules. I suspect most consumers will
be able to adopt the new major version without having to do anything in
practice.

This is intended to address:
https://github.com/w3c/webref/issues/1697