Add missing `descriptors` property to nested at-rules
The nested at-rules get extracted as descriptors and do not go through the
usual logic that handles at-rules. That's fine because they don't (currently at
least) define descriptors themselves. But that means they don't have a
`descriptors` property themselves. That property needs to be added when the
nested at-rules get moved to the root.

... and one day, I'll add schema testing in post-processing tests as well to
catch this sort of problems right away!