[Events] Simplify "not bubbling" logic and fix missing targets (#1984)
This fixes #1983. We had been erring on the safe side when consolidating events
by only flagging events that fire at (non root) interfaces in a tree as non
bubbling when the event did not fire at the root interface. That approach works
but there was a bug that made consolidation forget non root interfaces when an
event did not have a `bubbles` flag.

In practice, in the absence of explicit information, it seems fine to flag all
events that fire at non root interfaces in a tree as non bubbling. This update
simplifies the consolidation logic to always do that. That simplification fixes
the underlying bug as well.

Impacted events are: `blur`, `clipboardchange`, `error`, `focus`,
`gamepadconnected`, `gamepaddisconnected`, `load`, `securitypolicyviolation`
(the last event should actually bubble but the extraction logic cannot detect
that without some additional help).