[CSS] Avoid accidently looking at the entire body
Code looks at the surrounding context next to a definition to find nearby
production rules. When a definition is the direct child of the `<body>` or
`<main>` tag, it does not make sense to look for a production rule (we're
no longer "nearby").

This update catches the case to avoid accidental hiccups as recently found
in css-forms.